requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.7
flake8==6.0.0
mccabe==0.7.0
mypy==1.0.1
mypy-extensions==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
RPi.GPIO==0.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
